requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.16
fastapi-users[sqlalchemy]>=13.0.0
python-dotenv>=1.0.0
aiosmtplib>=3.0.0
aiosqlite>=0.20.0